Moved webserver to separate script
